feat: move controllers to internal/controller
